aurel debug$

+++ Common/inc/IServerPacket.hh
@@ -50,7 +50,7 @@ typedef struct
 typedef struct
 {
   Bool				success;
-  uint64_t			timestamp;
+  float				timestamp;
   uint32_t			magic;
 } AuthUDPData;
